+++ src/main/java/k5s/reviewdevelop/exception/NoMovieException.java
@@ -1,0 +1,7 @@
+package k5s.reviewdevelop.exception;
+
+public class NoMovieException extends InvalidAuthenticationException{
+    public NoMovieException(String message) {
+        super(message);
+    }
+}
