+++ src/main/resources/static/reviews/js/custom.js
@@ -1,0 +1,7 @@
+function cancel(reviewId) {
+        var form = document.createElement("form");
+        form.setAttribute("method", "post");
+        form.setAttribute("action", "/reviews/" + reviewId + "/cancel");
+        document.body.appendChild(form);
+        form.submit();
+    }
