+++ src/main/resources/templates/fragments/header.html
@@ -9,6 +9,7 @@
     <link rel="icon" type="image/png" href="https://k5s-member-convenience.s3.ap-northeast-2.amazonaws.com/common/k5s-icon.png" />
     <!--<link href="https://k5s-review.s3.ap-northeast-2.amazonaws.com/css/custom.css" rel="stylesheet">-->
     <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-1BmE4kWBq78iYhFldvKuhfTAU6auU8tT94WrHftjDbrCEXSU1oBoqyl2QvZ6jIW3" crossorigin="anonymous">
+    <script src="/reviews/js/custom.js" defer></script>
     <link href="/reviews/css/custom.css" rel="stylesheet">
     <link rel="preconnect" href="https://fonts.googleapis.com">
     <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
