+++ src/main/resources/templates/member/reviews/list.html
@@ -20,9 +20,9 @@
             <div class="card-body">
                 <h5 class="card-title" th:text="${review.description}">Success card title</h5>
                 <div class="text-end">
-                    <a href="#" th:href="@{/movies/{movieId}/reviews/{reviewId}/edit/my (movieId=${review.movie.id}, reviewId=${review.id})}" class="btn btn-dark" role="button">수정</a>
+                    <a href="#" th:href="@{/reviews/{reviewId}/edit/my (reviewId=${review.id})}" class="btn btn-dark" role="button">수정</a>
                     <a href="#"
-                       th:href="'javascript:cancel('+${review.movie.id}+', '+${review.id}+')'"
+                       th:href="'javascript:cancel('+${review.id}+')'"
                        class="btn btn-danger">삭제</a>
                 </div>
             </div>
@@ -32,13 +32,4 @@
     <div th:replace="fragments/footer :: footer"/>
 </div> <!-- /container -->
 </body>
-<script>
-    function cancel(movieId, reviewId) {
-        var form = document.createElement("form");
-        form.setAttribute("method", "post");
-        form.setAttribute("action", "/movies/" + movieId + "/reviews/" + reviewId + "/cancel");
-        document.body.appendChild(form);
-        form.submit();
-    }
-</script>
 </html>
