+++ src/main/resources/templates/member/reviews/shortList.html
@@ -14,9 +14,9 @@
                 <div class="card-body">
                     <h5 class="card-title" th:text="${review.description}">Success card title</h5>
                     <div class="text-end">
-                        <a href="#" th:href="@{/movies/{movieId}/reviews/{reviewId}/edit/short-my (movieId=${review.movie.id}, reviewId=${review.id})}" class="btn btn-dark" role="button">수정</a>
+                        <a href="#" th:href="@{/reviews/{reviewId}/edit/short-my (reviewId=${review.id})}" class="btn btn-dark" role="button">수정</a>
                         <a href="#"
-                           th:href="'javascript:cancel('+${review.movie.id}+', '+${review.id}+')'"
+                           th:href="'javascript:cancel('+${review.id}+')'"
                            class="btn btn-danger">삭제</a>
                     </div>
                 </div>
@@ -26,13 +26,4 @@
     </div>
 </div> <!-- /container -->
 </body>
-<script>
-    function cancel(movieId, reviewId) {
-        var form = document.createElement("form");
-        form.setAttribute("method", "post");
-        form.setAttribute("action", "/movies/" + movieId + "/reviews/" + reviewId + "/cancel");
-        document.body.appendChild(form);
-        form.submit();
-    }
-</script>
 </html>
