+++ src/main/resources/templates/movies/reviews/error.html
@@ -8,7 +8,7 @@
     <link href="//maxcdn.bootstrapcdn.com/bootstrap/4.1.1/css/bootstrap.min.css" rel="stylesheet" id="bootstrap-css">
     <script src="//maxcdn.bootstrapcdn.com/bootstrap/4.1.1/js/bootstrap.min.js"></script>
     <script src="//cdnjs.cloudflare.com/ajax/libs/jquery/3.2.1/jquery.min.js"></script>
-    <link href="/css/error.css" rel="stylesheet">
+    <link href="/reviews/css/error.css" rel="stylesheet">
 </head>
 <body>
 <div class="error-content">
